livekit-scribe/agent-server/agent/requirements.txt
livekit-agents>=1.4.4
livekit-plugins-deepgram>=0.6.0
livekit-api>=1.1.0
httpx>=0.27.0
